refactor(#19): clubServiceTest, scheduleServiceTest 예외처리 변경에 따른 코드 변경
