+++ src/test/java/com/example/moim/schedule/service/ScheduleServiceTest.java
@@ -3,9 +3,10 @@ package com.example.moim.schedule.service;
 import com.example.moim.club.dto.request.ClubInput;
 import com.example.moim.club.entity.Club;
 import com.example.moim.club.entity.UserClub;
-import com.example.moim.club.exception.ClubPermissionException;
+import com.example.moim.club.exception.advice.ClubControllerAdvice;
 import com.example.moim.club.repository.ClubRepository;
 import com.example.moim.club.repository.UserClubRepository;
+import com.example.moim.global.exception.ResponseCode;
 import com.example.moim.match.entity.Match;
 import com.example.moim.match.entity.MatchApplication;
 import com.example.moim.match.repository.MatchApplicationRepository;
@@ -14,6 +15,7 @@ import com.example.moim.notification.dto.ScheduleSaveEvent;
 import com.example.moim.schedule.dto.*;
 import com.example.moim.schedule.entity.Schedule;
 import com.example.moim.schedule.entity.ScheduleVote;
+import com.example.moim.schedule.exception.advice.ScheduleControllerAdvice;
 import com.example.moim.schedule.repository.ScheduleRepository;
 import com.example.moim.schedule.repository.ScheduleVoteRepository;
 import com.example.moim.user.dto.SignupInput;
@@ -117,9 +119,10 @@ class ScheduleServiceTest {
         when(clubRepository.findById(any(Long.class))).thenReturn(Optional.of(club));
         when(userClubRepository.findByClubAndUser(club, user)).thenReturn(Optional.of(userClub));
         //then
-        assertThrows(ClubPermissionException.class, () -> {
+        Exception exception = assertThrows(ScheduleControllerAdvice.class, () -> {
             scheduleService.saveSchedule(scheduleInput, user);
         });
+        assertThat(exception.getMessage()).isEqualTo(ResponseCode.CLUB_PERMISSION_DENIED.getMessage());
         verify(clubRepository, times(1)).findById(any(Long.class));
         verify(userClubRepository, times(1)).findByClubAndUser(any(Club.class), any(User.class));
     }
@@ -156,9 +159,10 @@ class ScheduleServiceTest {
         when(userClubRepository.findByClubAndUser(any(Club.class), any(User.class))).thenReturn(Optional.of(userClub));
         when(clubRepository.findById(any(Long.class))).thenReturn(Optional.of(club));
         //then
-        assertThrows(ClubPermissionException.class, () -> {
+        Exception exception = assertThrows(ScheduleControllerAdvice.class, () -> {
             scheduleService.updateSchedule(scheduleUpdateInput, user);
         });
+        assertThat(exception.getMessage()).isEqualTo(ResponseCode.CLUB_PERMISSION_DENIED.getMessage());
         verify(userClubRepository, times(1)).findByClubAndUser(any(Club.class), any(User.class));
         verify(clubRepository, times(1)).findById(any(Long.class));
     }
@@ -345,9 +349,10 @@ class ScheduleServiceTest {
         when(scheduleRepository.findById(any(Long.class))).thenReturn(Optional.of(schedule));
         when(userClubRepository.findByClubAndUser(any(Club.class), any(User.class))).thenReturn(Optional.of(userClub));
         //then
-        assertThrows(ClubPermissionException.class, () -> {
+        Exception exception = assertThrows(ScheduleControllerAdvice.class, () -> {
             scheduleService.closeSchedule(1L, user);
         });
+        assertThat(exception.getMessage()).isEqualTo(ResponseCode.CLUB_PERMISSION_DENIED.getMessage());
         verify(scheduleRepository, times(1)).findById(any(Long.class));
         verify(userClubRepository, times(1)).findByClubAndUser(any(Club.class), any(User.class));
     }
